(fix): Renamed "Services" to "Aggregates".
